Upgrade Angler Rate change
I added to the code to upgrade the Angler towers smallCatchRate. It is not complete in that it currently only handles the smallRate and only affects EXISTING towers, but once we decide on individual vs. systemwide upgrades, I will go through and change this code accordingly to be a little cleaner and more consistent.