Additional Obstacle Avoidance
Added the ability for us to add terrain types and penalty values to different types of terrain. By default there is no penalty besides a penalty for drifting closer to obstacles (since we do not want the fish to swim right on the coastline). If we want to add rapids or some other type of waer hazard, we can add it via layers in this code.