Added in Grid shifting
The grid can now be placed in locations besides 0,0,0 and operate correctly when a list of targets is provided. By setting the offsetX and offsetY for the individual grid, the units can be made to go to the proper location rather than a location based on the grid holder. In other words, if I have a grid located at -25,0,0 and a target somewhere in that grid, by setting the offsetX to +25, the unit will go to the actual intended location rather than 25 units to the left of the target