First File Cleaning Pass
This is the first pass of removing unnecessary files and moving them to a secondary repository for storage. I have left in a number of art assets that I want to consult Sam on as well as Vector Field Pathfinding elements until I can fully replace it (so as to not break the game).