Node Pathfinding Test
I made a test of the node pathfinding solution (which uses no algorithm). On a batch of 1000 fish it is more stable in frame rate, but has not been tested on the main map. I also made some early attempts to create a shrunk Columbia River map which has mostly proven difficult due to the many interwoven parts that don't seem to scale well.