Added in Nodes to the Dam
Added in Node Pathfinding up to the dam location in the new Columbia River Level Scene. A decision will have to be made about how to handle "stuck" fish as collisions with the dam are not strictly necessary if the fish are only given one chance to cross the dam (i.e. they can just die if they fail)