Finished Placing Nodes
Added the remaining nodes and spawning grounds for the Salmon to run the whole new Columbia River Scene. I changed to the fish script so that the Swim function is simply a constant Update function. This seems to reduce some of the performance cost from the FixedUpdate calls across all the scene objects. Also added in a simple skybox to replace the default.