Fishermen to Angler Changes
Switched out the name of FishermanTower to AnglerTower. Replaced the references to these towers in other scripts (such as the Ranger) to refer to the new name. Also, modified the Angler prefab animation cycle to fix bug where anglers would be in the ground until triggered by  a fish. This will likely require a deeper animator change alongside the Ranger to include a more nuanced Default state animation.