Fish Movement Temporary Build Fix
Adjusted the nature of the pause button to work more appropriately for the new movement system (no need to cache and restore velocities -> prevents fish from being sent off course or airborne). This means that pausing now sets the timeScale to 0 which also halts all camera movement. The upcoming camera rework is aimed at solving this, but for the moment this can demonstrate the new movement and camera ideas in a demo.