Detached Camera Movement from Time
The camera movement now uses Time.unscaledDeltaTime, so that the pausing of the game does not affect the movement of the camera. If we decide to make a pause menu in the future, a condition can be created to halt camera movement with a preemptive return in the CameraController script.