Main Menu and Adjustments
Created a new scene to act as the Main Menu. This will include buttons to the level selection screens, options menu, tutorial level, and exit button. The color and alpha level of the upper water layer was also adjusted to try and allow salmon color to shine through more clearly. The tower panel was widened to take up the whole screen width up to the help button.