Pause Menu Buttons First Pass
This push contains the first attempt at each of the pause menu restart buttons. In particular, the Restart Turn button is not complete, but has some of its basic functions like removing towers from turns after the desired turn restart and removing the active fish. It still needs to have proper game state transitions added and the generations of salmon need to be tracked via a list to properly respawn them.