Ranger Tower Refactor
This push contains the refactor for the Ranger tower. The main changes are the removal of a "flash" effect on the Anglers affected by the Ranger that  did not work correctly as well as changing the height of the end position of the line drawn between the Ranger and affected Anglers. Some comments and notes were added to clarify some confusing operations.