Dam and Ladder Refactor
This push focuses on restructuring the place Dams and Ladders have in relation to Towers. This push has Dams and Ladders derive from TowerBase and removes FilterBase (due to it not actually being necessary after implementation of the TowerBase scripts). Also removed a few more elements that were not being used in the Ranger and Sealion scripts.