Angler and Ranger Hats now have different colored hats
So I had to change the Shader that the hats are created with in order to get this to work. Also the Salmon Run logo on the ranger hats change as well. I don't think I can stop that from happening without separating the hat and the logo into two different objects.